requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
pydantic[email]
python-multipart
sqlalchemy_imageattach
fastapi-mail
python-dotenv
pyjwt